+++ CHANGELOG.md
@@ -7,6 +7,13 @@ This project started near the end of the 2023, due my necessity to improve my ab
 I tried - same time ago - to use electron, but at last eventually I decided to try it again in python, and I am seeing good results in terms of lightness and practicality of development.
 
 # Versions
+## pengui-0.7.10
+* added --pendrive compression to produce;
+* added --links to produce;
+
+## pengui-0.7.9
+* extended check for packages eggs and package penguins-eggs (the testing version).
+
 ## pengui-0.7.8
 * included new icons from Charlie Martinez [Quirinux](https://www.quirinux.org/).